+++ src/main/java/com/shoping/book_my_product/service/CartService.java
@@ -5,12 +5,12 @@ import java.util.List;
 import com.shoping.book_my_product.entity.Cart;
 
 public interface CartService {
-	public Cart saveCart(long productId,long UserId);
+	public Cart saveCart(Integer productId,Integer UserId);
 	
-	public List<Cart> getCartByUser(long userId);
+	public List<Cart> getCartByUser(Integer userId);
 	
-	public long getCountCart(long userId);
+	public long getCountCart(Integer userId);
 
-	public void updateQuantity(String sym, long cid);
+	public void updateQuantity(String sym, Integer cid);
 	
 }
